+++ db/migrate/20140809125800_create_entries.rb
@@ -1,0 +1,10 @@
+class CreateEntries < ActiveRecord::Migration
+  def change
+    create_table :entries do |t|
+      t.string :name
+      t.text :note
+      t.references :category
+      t.references :friend
+    end
+  end
+end